feat(agents): created a package called agents
