+++ packages/agents/package.json
@@ -1,0 +1,27 @@
+{
+	"name": "@openzosma/agents",
+	"version": "0.1.0",
+	"private": true,
+	"type": "module",
+	"main": "dist/index.js",
+	"types": "dist/index.d.ts",
+	"exports": {
+		".": {
+			"types": "./dist/index.d.ts",
+			"import": "./dist/index.js"
+		}
+	},
+	"scripts": {
+		"build": "tsc",
+		"check": "tsc --noEmit"
+	},
+	"dependencies": {
+		"@mariozechner/pi-agent-core": "^0.60.0",
+		"@mariozechner/pi-ai": "^0.60.0",
+		"@mariozechner/pi-coding-agent": "^0.60.0"
+	},
+	"devDependencies": {
+		"@types/node": "^22.15.2",
+		"typescript": "^5.7.3"
+	}
+}
